fix app updates when there is a built image
